webserver/requirements.txt
amqp==5.0.3
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.16.59
botocore==1.19.59
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.3
django-bootstrap-form==3.4
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
jmespath==0.10.0
kombu==5.0.2
prompt-toolkit==3.0.14
pycurl==7.43.0.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.4
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
